fixed text and added stuff for prices, including scripts and text

+++ Assets/Scripts/Bank.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Bank : MonoBehaviour
 {
-    public int gold = 0;
+    public double gold = 0;
     public int swordValue = 50;
     public int totalClicks = 0;
